Fix the Travis build
Unfortunately we need some hacks here: tsdl 0.9.2 depends on SDL 2.0.5,
which isn't available on Travis, so we go out of our way to pin 0.9.1.

Also, newer versions of tsdl aren't compatible with 4.01, so we drop it
from the build now.

This could use some caching, but just getting things working is a step
in the right direction.